ALGO-11 Moved class Stack to the utils module
